most popular data fixed

+++ SD.Shared/Models/Popular/PopularScraping.cs
@@ -18,8 +18,8 @@
         public List<Thumbnail>? thumbnails { get; set; }
         public string? trailer { get; set; }
         public string? contentRating { get; set; }
-        public int? startYear { get; set; }
-        public int? endYear { get; set; }
+        public long? startYear { get; set; }
+        public long? endYear { get; set; }
         public string? releaseDate { get; set; }
         public List<string>? interests { get; set; }
         public List<string>? countriesOfOrigin { get; set; }
@@ -27,20 +27,20 @@
         public List<string>? spokenLanguages { get; set; }
         public List<string>? filmingLocations { get; set; }
         public List<ProductionCompany>? productionCompanies { get; set; }
-        public int? budget { get; set; }
+        public long? budget { get; set; }
         public long? grossWorldwide { get; set; }
         public List<string>? genres { get; set; }
         public bool isAdult { get; set; }
-        public int? runtimeMinutes { get; set; }
+        public long? runtimeMinutes { get; set; }
         public double? averageRating { get; set; }
-        public int? numVotes { get; set; }
-        public int? metascore { get; set; }
+        public long? numVotes { get; set; }
+        public long? metascore { get; set; }
     }
 
     public class Thumbnail
     {
         public string? url { get; set; }
-        public int? width { get; set; }
-        public int? height { get; set; }
+        public long? width { get; set; }
+        public long? height { get; set; }
     }
 }
